# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Сервис анализа патентной активности компаний.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/persons": {
      "get": {
        "tags": [
          "Persons"
        ],
        "summary": "Patents Stats",
        "description": "Получение статистики по патентам в разрезе различных категорий.\n\nВозвращает статистику распределения патентов по следующим категориям:\n* Топ-5 кодов ОКОПФ (Организационно-правовая форма) + прочие\n* Топ-5 кодов ОКВЭД (Вид экономической деятельности) + прочие\n* Топ-5 подкатегорий МПК (Международная патентная классификация) + прочие",
        "operationId": "patents_stats_persons_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonsList"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Persons"
        ],
        "summary": "Create Person",
        "description": "Создание новой записи о лице.\n\nParameters:\n    person (PersonCreate): Объект с данными о лице, содержащий следующие поля:\n        * kind (int): Вид лица (1 - Юридическое лицо, 2 - ИП, 3 - Физическое лицо)\n        * tax_number (str): ИНН\n        * full_name (str, опционально): Полное наименование\n        * short_name (str, опционально): Сокращенное наименование\n        * legal_address (str, опционально): Юридический адрес\n        * fact_address (str, опционально): Фактический адрес\n        * reg_date (date, опционально): Дата регистрации\n        * active (bool): Статус активности\n        * okopf (str, опционально): Код ОКОПФ\n        * okvad (str, опционально): Код ОКВЭД\n        * region (str, опционально): Регион регистрации\n        * uk (int): Признак участника кластера (1 - участник, 0 - не участник)\n        * support_type (str, опционально): Тип поддержки\n\nReturns:\n    PersonDB: Созданный объект лица",
        "operationId": "create_person_persons_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PersonCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonDB"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/persons/all_stats": {
      "get": {
        "tags": [
          "Persons"
        ],
        "summary": "Get Persons All Stats",
        "description": "Получение общей статистики по всем лицам.\n\nParameters:\n    filter_id (int, опционально): Идентификатор фильтра для списка ИНН.",
        "operationId": "get_persons_all_stats_persons_all_stats_get",
        "parameters": [
          {
            "name": "filter_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Filter Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonsAllStats"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/persons/msk_stats": {
      "get": {
        "tags": [
          "Persons"
        ],
        "summary": "Get Persons Msk Stats",
        "description": "Получение статистики по лицам, зарегистрированным в Москве.\n\nArgs:\n    filter_id (Optional[int]): опциональный идентификатор загруженного фильтра по списку ИНН.\n     session (AsyncSession): асинхронная сессия базы данных.\n\nReturns:\n    PersonsStats: словарь со статистикой.",
        "operationId": "get_persons_msk_stats_persons_msk_stats_get",
        "parameters": [
          {
            "name": "filter_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Filter Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonsMskStats"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/persons/{person_tax_number}": {
      "get": {
        "tags": [
          "Persons"
        ],
        "summary": "Get Person",
        "description": "Получение подробной информации о лице по ИНН.\n\nParameters:\n    person_tax_number (str): ИНН лица\n\nReturns:\n    Объект PersonAdditionalFields, содержащий:\n        * Все базовые поля лица\n        * category (str): Категория лица\n        * patents (list): Список связанных патентов с их видами и регистрационными номерами\n        * patent_count (int): Общее количество патентов",
        "operationId": "get_person_persons__person_tax_number__get",
        "parameters": [
          {
            "name": "person_tax_number",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Person Tax Number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonAdditionalFields"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Persons"
        ],
        "summary": "Update Person",
        "description": "Обновление информации о существующем лице.\n\nArgs:\n    person_tax_number (str): идентификационный номер персоны.\n    obj_in (PersonUpdate): данные для обновления персоны.\n    session (AsyncSession): асинхронная сессия базы данных.\n\nReturns:\n    PersonDB: обновленная персона.",
        "operationId": "update_person_persons__person_tax_number__patch",
        "parameters": [
          {
            "name": "person_tax_number",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Person Tax Number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PersonUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonDB"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Persons"
        ],
        "summary": "Delete Person",
        "description": "Удаление лица по ИНН.\n\nArgs:\n    person_tax_number(int): идентификационный номер персоны.\n    session (AsyncSession): асинхронная сессия базы данных.",
        "operationId": "delete_person_persons__person_tax_number__delete",
        "parameters": [
          {
            "name": "person_tax_number",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Person Tax Number"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PersonAdditionalFields": {
        "properties": {
          "kind": {
            "type": "integer",
            "title": "Kind"
          },
          "tax_number": {
            "type": "string",
            "title": "Tax Number"
          },
          "full_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Full Name"
          },
          "short_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Short Name"
          },
          "legal_address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Legal Address"
          },
          "fact_address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fact Address"
          },
          "reg_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reg Date"
          },
          "active": {
            "type": "boolean",
            "title": "Active"
          },
          "okopf": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Okopf"
          },
          "okvad": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Okvad"
          },
          "region": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Region"
          },
          "uk": {
            "type": "integer",
            "title": "Uk"
          },
          "support_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Support Type"
          },
          "category": {
            "type": "string",
            "title": "Category"
          },
          "patents": {
            "items": {
              "$ref": "#/components/schemas/PersonPatents"
            },
            "type": "array",
            "title": "Patents",
            "default": []
          },
          "patent_count": {
            "type": "integer",
            "title": "Patent Count",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "kind",
          "tax_number",
          "full_name",
          "short_name",
          "legal_address",
          "fact_address",
          "reg_date",
          "active",
          "okopf",
          "okvad",
          "region",
          "uk",
          "support_type",
          "category"
        ],
        "title": "PersonAdditionalFields"
      },
      "PersonCreate": {
        "properties": {
          "kind": {
            "type": "integer",
            "title": "Kind"
          },
          "tax_number": {
            "type": "string",
            "title": "Tax Number"
          },
          "full_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Full Name"
          },
          "short_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Short Name"
          },
          "legal_address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Legal Address"
          },
          "fact_address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fact Address"
          },
          "reg_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reg Date"
          },
          "active": {
            "type": "boolean",
            "title": "Active"
          },
          "okopf": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Okopf"
          },
          "okvad": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Okvad"
          },
          "region": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Region"
          },
          "uk": {
            "type": "integer",
            "title": "Uk"
          },
          "support_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Support Type"
          }
        },
        "type": "object",
        "required": [
          "kind",
          "tax_number",
          "full_name",
          "short_name",
          "legal_address",
          "fact_address",
          "reg_date",
          "active",
          "okopf",
          "okvad",
          "region",
          "uk",
          "support_type"
        ],
        "title": "PersonCreate"
      },
      "PersonDB": {
        "properties": {
          "kind": {
            "type": "integer",
            "title": "Kind"
          },
          "tax_number": {
            "type": "string",
            "title": "Tax Number"
          },
          "full_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Full Name"
          },
          "short_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Short Name"
          },
          "legal_address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Legal Address"
          },
          "fact_address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fact Address"
          },
          "reg_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reg Date"
          },
          "active": {
            "type": "boolean",
            "title": "Active"
          },
          "okopf": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Okopf"
          },
          "okvad": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Okvad"
          },
          "region": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Region"
          },
          "uk": {
            "type": "integer",
            "title": "Uk"
          },
          "support_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Support Type"
          }
        },
        "type": "object",
        "required": [
          "kind",
          "tax_number",
          "full_name",
          "short_name",
          "legal_address",
          "fact_address",
          "reg_date",
          "active",
          "okopf",
          "okvad",
          "region",
          "uk",
          "support_type"
        ],
        "title": "PersonDB"
      },
      "PersonPatents": {
        "properties": {
          "kind": {
            "type": "integer",
            "title": "Kind"
          },
          "reg_number": {
            "type": "integer",
            "title": "Reg Number"
          }
        },
        "type": "object",
        "required": [
          "kind",
          "reg_number"
        ],
        "title": "PersonPatents"
      },
      "PersonUpdate": {
        "properties": {
          "kind": {
            "type": "integer",
            "title": "Kind"
          },
          "tax_number": {
            "type": "string",
            "title": "Tax Number"
          },
          "full_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Full Name"
          },
          "short_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Short Name"
          },
          "legal_address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Legal Address"
          },
          "fact_address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fact Address"
          },
          "reg_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reg Date"
          },
          "active": {
            "type": "boolean",
            "title": "Active"
          },
          "okopf": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Okopf"
          },
          "okvad": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Okvad"
          },
          "region": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Region"
          },
          "uk": {
            "type": "integer",
            "title": "Uk"
          },
          "support_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Support Type"
          }
        },
        "type": "object",
        "required": [
          "kind",
          "tax_number",
          "full_name",
          "short_name",
          "legal_address",
          "fact_address",
          "reg_date",
          "active",
          "okopf",
          "okvad",
          "region",
          "uk",
          "support_type"
        ],
        "title": "PersonUpdate"
      },
      "PersonsAllStats": {
        "properties": {
          "total_persons": {
            "type": "integer",
            "title": "Total Persons"
          },
          "by_kind": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "By Kind"
          },
          "by_category": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "By Category"
          }
        },
        "type": "object",
        "required": [
          "total_persons",
          "by_kind",
          "by_category"
        ],
        "title": "PersonsAllStats"
      },
      "PersonsList": {
        "properties": {
          "okopf_stats": {
            "items": {
              "$ref": "#/components/schemas/StatItem"
            },
            "type": "array",
            "title": "Okopf Stats"
          },
          "okvad_stats": {
            "items": {
              "$ref": "#/components/schemas/StatItem"
            },
            "type": "array",
            "title": "Okvad Stats"
          },
          "mpk_stats": {
            "items": {
              "$ref": "#/components/schemas/StatItem"
            },
            "type": "array",
            "title": "Mpk Stats"
          }
        },
        "type": "object",
        "required": [
          "okopf_stats",
          "okvad_stats",
          "mpk_stats"
        ],
        "title": "PersonsList"
      },
      "PersonsMskStats": {
        "properties": {
          "total_persons": {
            "type": "integer",
            "title": "Total Persons"
          },
          "by_kind": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "By Kind"
          },
          "by_category": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "By Category"
          },
          "moscow_cluster_percentage": {
            "type": "number",
            "title": "Moscow Cluster Percentage"
          },
          "moscow_support_type_percentage": {
            "type": "number",
            "title": "Moscow Support Type Percentage"
          }
        },
        "type": "object",
        "required": [
          "total_persons",
          "by_kind",
          "by_category",
          "moscow_cluster_percentage",
          "moscow_support_type_percentage"
        ],
        "title": "PersonsMskStats"
      },
      "StatItem": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          }
        },
        "type": "object",
        "required": [
          "name",
          "count"
        ],
        "title": "StatItem"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}